+++ instrumentation/core/src/main/java/de/mannodermaus/junit5/condition/EnabledOnManufacturer.kt
@@ -1,8 +1,12 @@
 package de.mannodermaus.junit5.condition
 
+import de.mannodermaus.junit5.internal.EnabledOnManufacturerCondition
 import org.junit.jupiter.api.extension.ExtendWith
 
 @Target(AnnotationTarget.CLASS, AnnotationTarget.FUNCTION)
 @Retention(AnnotationRetention.RUNTIME)
 @ExtendWith(EnabledOnManufacturerCondition::class)
-annotation class EnabledOnManufacturer(val value: Array<String>, val ignoreCase: Boolean = true)
+public annotation class EnabledOnManufacturer(
+    val value: Array<String>,
+    val ignoreCase: Boolean = true
+)