+++ instrumentation/core/src/main/java/de/mannodermaus/junit5/internal/CoreInternalConstants.kt
@@ -1,0 +1,3 @@
+package de.mannodermaus.junit5.internal
+
+internal const val NOT_SET = -1