+++ instrumentation/core/src/test/java/de/mannodermaus/junit5/condition/EnabledOnSdkVersionConditionTests.kt
@@ -1,6 +1,7 @@
 package de.mannodermaus.junit5.condition
 
 import com.google.common.truth.Truth.assertThat
+import de.mannodermaus.junit5.internal.EnabledOnSdkVersionCondition
 import de.mannodermaus.junit5.util.AndroidBuildUtils.withApiLevel
 import org.junit.jupiter.api.Test
 import org.junit.jupiter.api.assertThrows