+++ instrumentation/runner/src/main/kotlin/de/mannodermaus/junit5/internal/discovery/PropertiesParser.kt
@@ -1,0 +1,10 @@
+package de.mannodermaus.junit5.internal.discovery
+
+internal object PropertiesParser {
+    @JvmStatic
+    fun fromString(string: String) =
+        string.split(",")
+            .map { keyValuePair -> keyValuePair.split("=") }
+            .filter { keyValueList -> keyValueList.size == 2 }
+            .associate { it[0] to it[1] }
+}